// src/utils/enum/Message.enum.tsx
enum Message {
  called = 'Ligou',
  hangUp = 'Desligou!',
  connected = 'Já está ligada!',
  alreadyOff = 'Já está desligada!',
  empty = 'Primeiro adicione um número',
  ivalid = 'Expressão inválida',
  errorCalc = 'Erro no cálculo',
  errorExpression = 'Erro na expressão',
  info = 'Clique no botão AC para ligar a calculadora',
}

export default Message;